Adding new configuration option: KeepDownloads. When is set to true, donwloads are not deleted after extract.

+++ Unverum/ModDownloader.cs
@@ -217,7 +217,8 @@ namespace Unverum
                 else
                 {
                     // Only delete if successfully extracted
-                    File.Delete(_ArchiveSource);
+                    if (!Global.config.KeepDownloads)
+                        File.Delete(_ArchiveSource);
                 }
             });
 
@@ -310,7 +311,8 @@ namespace Unverum
                 else
                 {
                     // Only delete if successfully extracted
-                    File.Delete(_ArchiveSource);
+                    if (!Global.config.KeepDownloads)
+                        File.Delete(_ArchiveSource);
                 }
             });
